Upgrade tanstack start version
