--- apps/auth/app/client-entry.tsx
@@ -1,14 +1,0 @@
-/// <reference types="vinxi/types/client" />
-import { StartClient } from '@tanstack/start';
-import { StrictMode } from 'react';
-import { hydrateRoot } from 'react-dom/client';
-import { createRouter } from './router';
-
-const router = createRouter();
-
-hydrateRoot(
-  document,
-  <StrictMode>
-    <StartClient router={router} />
-  </StrictMode>,
-);
