--- apps/auth/app/route-tree.gen.ts
@@ -1,274 +1,0 @@
-/* eslint-disable */
-
-// @ts-nocheck
-
-// noinspection JSUnusedGlobalSymbols
-
-// This file was automatically generated by TanStack Router.
-// You should NOT make any changes in this file as it will be overwritten.
-// Additionally, you should also exclude this file from your linter and/or formatter to prevent it from being checked or modified.
-
-// Import Routes
-
-import { Route as rootRoute } from './routes/__root';
-import { Route as VerifyEmailImport } from './routes/verify-email';
-import { Route as TotpMfaImport } from './routes/totp-mfa';
-import { Route as ResetPasswordImport } from './routes/reset-password';
-import { Route as RegisterImport } from './routes/register';
-import { Route as LoginImport } from './routes/login';
-import { Route as EmailMfaImport } from './routes/email-mfa';
-import { Route as ChooseMfaImport } from './routes/choose-mfa';
-import { Route as IndexImport } from './routes/index';
-
-// Create/Update Routes
-
-const VerifyEmailRoute = VerifyEmailImport.update({
-  id: '/verify-email',
-  path: '/verify-email',
-  getParentRoute: () => rootRoute,
-} as any);
-
-const TotpMfaRoute = TotpMfaImport.update({
-  id: '/totp-mfa',
-  path: '/totp-mfa',
-  getParentRoute: () => rootRoute,
-} as any);
-
-const ResetPasswordRoute = ResetPasswordImport.update({
-  id: '/reset-password',
-  path: '/reset-password',
-  getParentRoute: () => rootRoute,
-} as any);
-
-const RegisterRoute = RegisterImport.update({
-  id: '/register',
-  path: '/register',
-  getParentRoute: () => rootRoute,
-} as any);
-
-const LoginRoute = LoginImport.update({
-  id: '/login',
-  path: '/login',
-  getParentRoute: () => rootRoute,
-} as any);
-
-const EmailMfaRoute = EmailMfaImport.update({
-  id: '/email-mfa',
-  path: '/email-mfa',
-  getParentRoute: () => rootRoute,
-} as any);
-
-const ChooseMfaRoute = ChooseMfaImport.update({
-  id: '/choose-mfa',
-  path: '/choose-mfa',
-  getParentRoute: () => rootRoute,
-} as any);
-
-const IndexRoute = IndexImport.update({
-  id: '/',
-  path: '/',
-  getParentRoute: () => rootRoute,
-} as any);
-
-// Populate the FileRoutesByPath interface
-
-declare module '@tanstack/react-router' {
-  interface FileRoutesByPath {
-    '/': {
-      id: '/';
-      path: '/';
-      fullPath: '/';
-      preLoaderRoute: typeof IndexImport;
-      parentRoute: typeof rootRoute;
-    };
-    '/choose-mfa': {
-      id: '/choose-mfa';
-      path: '/choose-mfa';
-      fullPath: '/choose-mfa';
-      preLoaderRoute: typeof ChooseMfaImport;
-      parentRoute: typeof rootRoute;
-    };
-    '/email-mfa': {
-      id: '/email-mfa';
-      path: '/email-mfa';
-      fullPath: '/email-mfa';
-      preLoaderRoute: typeof EmailMfaImport;
-      parentRoute: typeof rootRoute;
-    };
-    '/login': {
-      id: '/login';
-      path: '/login';
-      fullPath: '/login';
-      preLoaderRoute: typeof LoginImport;
-      parentRoute: typeof rootRoute;
-    };
-    '/register': {
-      id: '/register';
-      path: '/register';
-      fullPath: '/register';
-      preLoaderRoute: typeof RegisterImport;
-      parentRoute: typeof rootRoute;
-    };
-    '/reset-password': {
-      id: '/reset-password';
-      path: '/reset-password';
-      fullPath: '/reset-password';
-      preLoaderRoute: typeof ResetPasswordImport;
-      parentRoute: typeof rootRoute;
-    };
-    '/totp-mfa': {
-      id: '/totp-mfa';
-      path: '/totp-mfa';
-      fullPath: '/totp-mfa';
-      preLoaderRoute: typeof TotpMfaImport;
-      parentRoute: typeof rootRoute;
-    };
-    '/verify-email': {
-      id: '/verify-email';
-      path: '/verify-email';
-      fullPath: '/verify-email';
-      preLoaderRoute: typeof VerifyEmailImport;
-      parentRoute: typeof rootRoute;
-    };
-  }
-}
-
-// Create and export the route tree
-
-export interface FileRoutesByFullPath {
-  '/': typeof IndexRoute;
-  '/choose-mfa': typeof ChooseMfaRoute;
-  '/email-mfa': typeof EmailMfaRoute;
-  '/login': typeof LoginRoute;
-  '/register': typeof RegisterRoute;
-  '/reset-password': typeof ResetPasswordRoute;
-  '/totp-mfa': typeof TotpMfaRoute;
-  '/verify-email': typeof VerifyEmailRoute;
-}
-
-export interface FileRoutesByTo {
-  '/': typeof IndexRoute;
-  '/choose-mfa': typeof ChooseMfaRoute;
-  '/email-mfa': typeof EmailMfaRoute;
-  '/login': typeof LoginRoute;
-  '/register': typeof RegisterRoute;
-  '/reset-password': typeof ResetPasswordRoute;
-  '/totp-mfa': typeof TotpMfaRoute;
-  '/verify-email': typeof VerifyEmailRoute;
-}
-
-export interface FileRoutesById {
-  __root__: typeof rootRoute;
-  '/': typeof IndexRoute;
-  '/choose-mfa': typeof ChooseMfaRoute;
-  '/email-mfa': typeof EmailMfaRoute;
-  '/login': typeof LoginRoute;
-  '/register': typeof RegisterRoute;
-  '/reset-password': typeof ResetPasswordRoute;
-  '/totp-mfa': typeof TotpMfaRoute;
-  '/verify-email': typeof VerifyEmailRoute;
-}
-
-export interface FileRouteTypes {
-  fileRoutesByFullPath: FileRoutesByFullPath;
-  fullPaths:
-    | '/'
-    | '/choose-mfa'
-    | '/email-mfa'
-    | '/login'
-    | '/register'
-    | '/reset-password'
-    | '/totp-mfa'
-    | '/verify-email';
-  fileRoutesByTo: FileRoutesByTo;
-  to:
-    | '/'
-    | '/choose-mfa'
-    | '/email-mfa'
-    | '/login'
-    | '/register'
-    | '/reset-password'
-    | '/totp-mfa'
-    | '/verify-email';
-  id:
-    | '__root__'
-    | '/'
-    | '/choose-mfa'
-    | '/email-mfa'
-    | '/login'
-    | '/register'
-    | '/reset-password'
-    | '/totp-mfa'
-    | '/verify-email';
-  fileRoutesById: FileRoutesById;
-}
-
-export interface RootRouteChildren {
-  IndexRoute: typeof IndexRoute;
-  ChooseMfaRoute: typeof ChooseMfaRoute;
-  EmailMfaRoute: typeof EmailMfaRoute;
-  LoginRoute: typeof LoginRoute;
-  RegisterRoute: typeof RegisterRoute;
-  ResetPasswordRoute: typeof ResetPasswordRoute;
-  TotpMfaRoute: typeof TotpMfaRoute;
-  VerifyEmailRoute: typeof VerifyEmailRoute;
-}
-
-const rootRouteChildren: RootRouteChildren = {
-  IndexRoute: IndexRoute,
-  ChooseMfaRoute: ChooseMfaRoute,
-  EmailMfaRoute: EmailMfaRoute,
-  LoginRoute: LoginRoute,
-  RegisterRoute: RegisterRoute,
-  ResetPasswordRoute: ResetPasswordRoute,
-  TotpMfaRoute: TotpMfaRoute,
-  VerifyEmailRoute: VerifyEmailRoute,
-};
-
-export const routeTree = rootRoute
-  ._addFileChildren(rootRouteChildren)
-  ._addFileTypes<FileRouteTypes>();
-
-/* ROUTE_MANIFEST_START
-{
-  "routes": {
-    "__root__": {
-      "filePath": "__root.tsx",
-      "children": [
-        "/",
-        "/choose-mfa",
-        "/email-mfa",
-        "/login",
-        "/register",
-        "/reset-password",
-        "/totp-mfa",
-        "/verify-email"
-      ]
-    },
-    "/": {
-      "filePath": "index.tsx"
-    },
-    "/choose-mfa": {
-      "filePath": "choose-mfa.tsx"
-    },
-    "/email-mfa": {
-      "filePath": "email-mfa.tsx"
-    },
-    "/login": {
-      "filePath": "login.tsx"
-    },
-    "/register": {
-      "filePath": "register.tsx"
-    },
-    "/reset-password": {
-      "filePath": "reset-password.tsx"
-    },
-    "/totp-mfa": {
-      "filePath": "totp-mfa.tsx"
-    },
-    "/verify-email": {
-      "filePath": "verify-email.tsx"
-    }
-  }
-}
-ROUTE_MANIFEST_END */
