--- apps/auth/app/ssr-entry.tsx
@@ -1,9 +1,0 @@
-import { getRouterManifest } from '@tanstack/start/router-manifest';
-import { createStartHandler, defaultStreamHandler } from '@tanstack/start/server';
-
-import { createRouter } from './router';
-
-export default createStartHandler({
-  createRouter,
-  getRouterManifest,
-})(defaultStreamHandler);
