+++ apps/auth/src/components/not-found.tsx
