+++ apps/auth/src/route-tree.gen.ts
@@ -1,0 +1,210 @@
+/* eslint-disable */
+
+// @ts-nocheck
+
+// noinspection JSUnusedGlobalSymbols
+
+// This file was automatically generated by TanStack Router.
+// You should NOT make any changes in this file as it will be overwritten.
+// Additionally, you should also exclude this file from your linter and/or formatter to prevent it from being checked or modified.
+
+import { Route as rootRouteImport } from './routes/__root';
+import { Route as VerifyEmailRouteImport } from './routes/verify-email';
+import { Route as TotpMfaRouteImport } from './routes/totp-mfa';
+import { Route as ResetPasswordRouteImport } from './routes/reset-password';
+import { Route as RegisterRouteImport } from './routes/register';
+import { Route as LoginRouteImport } from './routes/login';
+import { Route as EmailMfaRouteImport } from './routes/email-mfa';
+import { Route as ChooseMfaRouteImport } from './routes/choose-mfa';
+import { Route as IndexRouteImport } from './routes/index';
+
+const VerifyEmailRoute = VerifyEmailRouteImport.update({
+  id: '/verify-email',
+  path: '/verify-email',
+  getParentRoute: () => rootRouteImport,
+} as any);
+const TotpMfaRoute = TotpMfaRouteImport.update({
+  id: '/totp-mfa',
+  path: '/totp-mfa',
+  getParentRoute: () => rootRouteImport,
+} as any);
+const ResetPasswordRoute = ResetPasswordRouteImport.update({
+  id: '/reset-password',
+  path: '/reset-password',
+  getParentRoute: () => rootRouteImport,
+} as any);
+const RegisterRoute = RegisterRouteImport.update({
+  id: '/register',
+  path: '/register',
+  getParentRoute: () => rootRouteImport,
+} as any);
+const LoginRoute = LoginRouteImport.update({
+  id: '/login',
+  path: '/login',
+  getParentRoute: () => rootRouteImport,
+} as any);
+const EmailMfaRoute = EmailMfaRouteImport.update({
+  id: '/email-mfa',
+  path: '/email-mfa',
+  getParentRoute: () => rootRouteImport,
+} as any);
+const ChooseMfaRoute = ChooseMfaRouteImport.update({
+  id: '/choose-mfa',
+  path: '/choose-mfa',
+  getParentRoute: () => rootRouteImport,
+} as any);
+const IndexRoute = IndexRouteImport.update({
+  id: '/',
+  path: '/',
+  getParentRoute: () => rootRouteImport,
+} as any);
+
+export interface FileRoutesByFullPath {
+  '/': typeof IndexRoute;
+  '/choose-mfa': typeof ChooseMfaRoute;
+  '/email-mfa': typeof EmailMfaRoute;
+  '/login': typeof LoginRoute;
+  '/register': typeof RegisterRoute;
+  '/reset-password': typeof ResetPasswordRoute;
+  '/totp-mfa': typeof TotpMfaRoute;
+  '/verify-email': typeof VerifyEmailRoute;
+}
+export interface FileRoutesByTo {
+  '/': typeof IndexRoute;
+  '/choose-mfa': typeof ChooseMfaRoute;
+  '/email-mfa': typeof EmailMfaRoute;
+  '/login': typeof LoginRoute;
+  '/register': typeof RegisterRoute;
+  '/reset-password': typeof ResetPasswordRoute;
+  '/totp-mfa': typeof TotpMfaRoute;
+  '/verify-email': typeof VerifyEmailRoute;
+}
+export interface FileRoutesById {
+  __root__: typeof rootRouteImport;
+  '/': typeof IndexRoute;
+  '/choose-mfa': typeof ChooseMfaRoute;
+  '/email-mfa': typeof EmailMfaRoute;
+  '/login': typeof LoginRoute;
+  '/register': typeof RegisterRoute;
+  '/reset-password': typeof ResetPasswordRoute;
+  '/totp-mfa': typeof TotpMfaRoute;
+  '/verify-email': typeof VerifyEmailRoute;
+}
+export interface FileRouteTypes {
+  fileRoutesByFullPath: FileRoutesByFullPath;
+  fullPaths:
+    | '/'
+    | '/choose-mfa'
+    | '/email-mfa'
+    | '/login'
+    | '/register'
+    | '/reset-password'
+    | '/totp-mfa'
+    | '/verify-email';
+  fileRoutesByTo: FileRoutesByTo;
+  to:
+    | '/'
+    | '/choose-mfa'
+    | '/email-mfa'
+    | '/login'
+    | '/register'
+    | '/reset-password'
+    | '/totp-mfa'
+    | '/verify-email';
+  id:
+    | '__root__'
+    | '/'
+    | '/choose-mfa'
+    | '/email-mfa'
+    | '/login'
+    | '/register'
+    | '/reset-password'
+    | '/totp-mfa'
+    | '/verify-email';
+  fileRoutesById: FileRoutesById;
+}
+export interface RootRouteChildren {
+  IndexRoute: typeof IndexRoute;
+  ChooseMfaRoute: typeof ChooseMfaRoute;
+  EmailMfaRoute: typeof EmailMfaRoute;
+  LoginRoute: typeof LoginRoute;
+  RegisterRoute: typeof RegisterRoute;
+  ResetPasswordRoute: typeof ResetPasswordRoute;
+  TotpMfaRoute: typeof TotpMfaRoute;
+  VerifyEmailRoute: typeof VerifyEmailRoute;
+}
+
+declare module '@tanstack/react-router' {
+  interface FileRoutesByPath {
+    '/verify-email': {
+      id: '/verify-email';
+      path: '/verify-email';
+      fullPath: '/verify-email';
+      preLoaderRoute: typeof VerifyEmailRouteImport;
+      parentRoute: typeof rootRouteImport;
+    };
+    '/totp-mfa': {
+      id: '/totp-mfa';
+      path: '/totp-mfa';
+      fullPath: '/totp-mfa';
+      preLoaderRoute: typeof TotpMfaRouteImport;
+      parentRoute: typeof rootRouteImport;
+    };
+    '/reset-password': {
+      id: '/reset-password';
+      path: '/reset-password';
+      fullPath: '/reset-password';
+      preLoaderRoute: typeof ResetPasswordRouteImport;
+      parentRoute: typeof rootRouteImport;
+    };
+    '/register': {
+      id: '/register';
+      path: '/register';
+      fullPath: '/register';
+      preLoaderRoute: typeof RegisterRouteImport;
+      parentRoute: typeof rootRouteImport;
+    };
+    '/login': {
+      id: '/login';
+      path: '/login';
+      fullPath: '/login';
+      preLoaderRoute: typeof LoginRouteImport;
+      parentRoute: typeof rootRouteImport;
+    };
+    '/email-mfa': {
+      id: '/email-mfa';
+      path: '/email-mfa';
+      fullPath: '/email-mfa';
+      preLoaderRoute: typeof EmailMfaRouteImport;
+      parentRoute: typeof rootRouteImport;
+    };
+    '/choose-mfa': {
+      id: '/choose-mfa';
+      path: '/choose-mfa';
+      fullPath: '/choose-mfa';
+      preLoaderRoute: typeof ChooseMfaRouteImport;
+      parentRoute: typeof rootRouteImport;
+    };
+    '/': {
+      id: '/';
+      path: '/';
+      fullPath: '/';
+      preLoaderRoute: typeof IndexRouteImport;
+      parentRoute: typeof rootRouteImport;
+    };
+  }
+}
+
+const rootRouteChildren: RootRouteChildren = {
+  IndexRoute: IndexRoute,
+  ChooseMfaRoute: ChooseMfaRoute,
+  EmailMfaRoute: EmailMfaRoute,
+  LoginRoute: LoginRoute,
+  RegisterRoute: RegisterRoute,
+  ResetPasswordRoute: ResetPasswordRoute,
+  TotpMfaRoute: TotpMfaRoute,
+  VerifyEmailRoute: VerifyEmailRoute,
+};
+export const routeTree = rootRouteImport
+  ._addFileChildren(rootRouteChildren)
+  ._addFileTypes<FileRouteTypes>();
