+++ apps/auth/src/routes/__root.tsx
@@ -1,26 +1,25 @@
+/// <reference types="vite/client" />
 import { publicEnv } from '@repo/env';
 import { ThemeProvider } from '@repo/ui/components/theme';
-// @ts-expect-error
-import styles from '@repo/ui/globals.css?url';
-import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
-import { createRootRoute, createRootRouteWithContext } from '@tanstack/react-router';
-import { Outlet, ScrollRestoration } from '@tanstack/react-router';
-import { Meta, Scripts } from '@tanstack/start';
+import { createRootRouteWithContext, HeadContent } from '@tanstack/react-router';
+import { Outlet, Scripts } from '@tanstack/react-router';
 import * as React from 'react';
 import NotFound from '~/components/not-found';
 import type { RouterContext } from '~/router';
+
+import styles from '@repo/ui/globals.css?url';
 
 export const Route = createRootRouteWithContext<RouterContext>()({
   head: () => ({
     meta: [
       { charSet: 'utf-8' },
       { name: 'viewport', content: 'width=device-width, initial-scale=1' },
-      { title: 'TrailEyes Panel' },
+      { title: 'TrailEyes Auth' },
     ],
     links: [{ rel: 'stylesheet', href: styles }],
   }),
   component: RootComponent,
-  notFoundComponent: () => <NotFound homepage="/" />,
+  notFoundComponent: () => <NotFound homepage="/register" />,
 });
 
 function RootComponent() {
@@ -31,14 +30,12 @@ function RootComponent() {
   );
 }
 
-const queryClient = new QueryClient();
-
 function RootDocument({ children }: React.PropsWithChildren) {
   const RouterDevtools =
     publicEnv().mode === 'production'
       ? () => null
       : React.lazy(() =>
-          import('@tanstack/router-devtools').then((mod) => ({
+          import('@tanstack/react-router-devtools').then((mod) => ({
             default: mod.TanStackRouterDevtools,
           })),
         );
@@ -55,7 +52,7 @@ function RootDocument({ children }: React.PropsWithChildren) {
   return (
     <html lang="en" suppressHydrationWarning>
       <head>
-        <Meta />
+        <HeadContent />
       </head>
       <body className="antialiased">
         <ThemeProvider>{children}</ThemeProvider>
@@ -63,7 +60,6 @@ function RootDocument({ children }: React.PropsWithChildren) {
           <RouterDevtools />
           <QueryDevtools />
         </React.Suspense>
-        <ScrollRestoration />
         <Scripts />
       </body>
     </html>
