+++ apps/auth/src/routes/register.tsx
