+++ apps/auth/src/util/external-redirect.ts
