+++ apps/auth/tsconfig.json
@@ -4,7 +4,7 @@
   "compilerOptions": {
     "baseUrl": ".",
     "paths": {
-      "~/*": ["app/*"]
+      "~/*": ["src/*"]
     }
   },
   "references": [
