+++ apps/backend/package.json
@@ -16,8 +16,8 @@
     "@repo/contract": "workspace:*",
     "zod": "^3.23.8",
     "hono": "^4.6.16",
-    "@orpc/server": "^1.6.3",
-    "@orpc/zod": "^1.6.3",
+    "@orpc/server": "^1.6.4",
+    "@orpc/zod": "^1.6.4",
     "lucia": "^3.2.0",
     "@paralleldrive/cuid2": "^2.2.2",
     "normalize-url": "^8.0.1",
