--- apps/panel/app/api-entry.ts
@@ -1,3 +1,0 @@
-import { createStartAPIHandler, defaultAPIFileRouteHandler } from '@tanstack/start/api';
-
-export default createStartAPIHandler(defaultAPIFileRouteHandler);
