--- apps/panel/app/client-entry.tsx
@@ -1,13 +1,0 @@
-import { StartClient } from '@tanstack/start';
-import { StrictMode } from 'react';
-import { hydrateRoot } from 'react-dom/client';
-import { createRouter } from './router';
-
-const router = createRouter();
-
-hydrateRoot(
-  document,
-  <StrictMode>
-    <StartClient router={router} />
-  </StrictMode>,
-);
