+++ apps/panel/package.json
@@ -2,37 +2,38 @@
   "name": "@repo/panel",
   "type": "module",
   "scripts": {
-    "dev": "vinxi dev --port 3001 --host",
-    "build": "vinxi build",
+    "dev": "vite dev --port 3001 --host",
+    "build": "vite build",
     "start": "PORT=3001 bun run .output/server/index.mjs"
   },
   "dependencies": {
     "@repo/ui": "workspace:*",
     "@repo/env": "workspace:*",
     "@repo/contract": "workspace:*",
-    "@ts-rest/core": "^3.51.0",
-    "@ts-rest/react-query": "^3.51.0",
-    "@tanstack/react-form": "^0.41.0",
-    "@tanstack/react-query": "^5.63.0",
-    "@tanstack/react-query-devtools": "^5.63.0",
-    "@tanstack/react-router": "^1.95.1",
-    "@tanstack/react-router-with-query": "^1.95.1",
-    "@tanstack/router-devtools": "^1.95.1",
-    "@tanstack/start": "^1.95.1",
-    "@vitejs/plugin-react": "^4.3.1",
-    "posthog-js": "^1.205.0",
-    "react": "^19.0.0",
-    "react-dom": "^19.0.0",
-    "vinxi": "^0.5.1",
-    "lucide-react": "^0.469.0",
+    "@orpc/tanstack-query": "^1.6.4",
+    "@orpc/client": "^1.6.4",
+    "@orpc/contract": "^1.6.4",
+    "@tanstack/react-form": "^1.12.4",
+    "@tanstack/react-query": "^5.81.5",
+    "@tanstack/react-query-devtools": "^5.81.5",
+    "@tanstack/react-router": "^1.124.0",
+    "@tanstack/react-router-with-query": "^1.124.0",
+    "@tanstack/react-router-devtools": "^1.124.0",
+    "@tanstack/react-start": "^1.124.1",
+    "@vitejs/plugin-react": "^4.6.0",
+    "posthog-js": "^1.256.2",
+    "react": "^19.1.0",
+    "react-dom": "^19.1.0",
+    "vite": "^7.0.2",
+    "lucide-react": "^0.525.0",
     "react-map-gl": "^7.1.8",
     "maplibre-gl": "^4.7.1",
     "vite-env-only": "^3.0.3"
   },
   "devDependencies": {
     "@repo/typescript-config": "workspace:*",
-    "@types/react": "^19.0.4",
-    "@types/react-dom": "^19.0.2",
+    "@types/react": "^19.1.8",
+    "@types/react-dom": "^19.1.6",
     "vite-tsconfig-paths": "^5.1.4",
     "autoprefixer": "^10.4.20",
     "postcss": "^8.4.49",
