+++ apps/panel/src/components/trail-eyes-map.css
