+++ apps/panel/src/routes/test.tsx
@@ -1,0 +1,9 @@
+import { createFileRoute } from '@tanstack/react-router';
+
+export const Route = createFileRoute('/test')({
+  component: RouteComponent,
+});
+
+function RouteComponent() {
+  return <div>Hello "/test"!</div>;
+}
