+++ packages/contract/package.json
@@ -10,8 +10,8 @@
     "@repo/database": "workspace:*",
     "@repo/zod-utils": "workspace:*",
     "zod": "^3.25.71",
-    "@orpc/contract": "^1.6.3",
-    "@orpc/zod": "^1.6.3"
+    "@orpc/contract": "^1.6.4",
+    "@orpc/zod": "^1.6.4"
   },
   "devDependencies": {
     "@types/bun": "latest",
