+++ packages/ui/package.json
@@ -27,20 +27,20 @@
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
     "input-otp": "^1.2.4",
-    "lucide-react": "^0.427.0",
-    "@tanstack/react-router": "^1.95.1",
-    "react": "^19.0.0",
+    "lucide-react": "^0.525.0",
+    "@tanstack/react-router": "^1.124.0",
+    "react": "^19.1.0",
     "react-hook-form": "^7.54.2",
     "tailwind-merge": "^2.5.2",
     "tailwindcss-animate": "^1.0.7",
-    "zod": "^3.23.8"
+    "zod": "^3.25.74"
   },
   "devDependencies": {
     "@repo/typescript-config": "workspace:*",
     "@turbo/gen": "^1.12.4",
     "@types/node": "^20.11.24",
-    "@types/react": "^19.0.4",
-    "@types/react-dom": "^19.0.2",
+    "@types/react": "^19.1.8",
+    "@types/react-dom": "^19.1.6",
     "typescript": "^5.3.3",
     "tailwindcss": "^3.4.10"
   }
